i fucking hated this so much
i was happy, happy as fucking fuck, until I realized in part 2 that my solution was wrong. I checked and checked and couldn't realize why, until I saw in the input something. `oneight`. I wondered "wait, is this supposed to be 1, 8 or 1?
after asking, the worst outcome came true. it was 1 8. at the time I was using regex, so I started looking for ways of somehow making the regex match overlaps. that took me probably an hour or two, and then after I found how it was supposed to be done (capture group lookahead), it worked in the sense of it detecting the overlapping numbers through an array, but it would return them as fucking null. i spent probably another hour trying to figure out why, and then I gave up on regex altogether.
another hour or two, and I made the monster that is part2's code. regex by fucking hand, iterating through each single fucking character by myself, i don't even care about trying to make this look pretty or polished or finished or optimized, I'm fucking done with this exercise. i am completing the rest of this advent because I am not a pussy, but MY FUCKING GOD if the rest of the exercises end up being this same shit, I am not touching it with a pole ever again.
the worst part is that not even the examples in the aoc page accounted for this, so they never made an example where you had something like `oneight`, that was a big mistake.
i am also not ever using javascript again on this advent, fuck you js.